requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
greenlet==1.1.2
idna==3.3
numpy==1.21.4
pandas==1.3.4
psycopg2-binary==2.9.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.3.24
urllib3==1.26.7
